+++ src/lib/decorators/serializable/index.ts
@@ -21,10 +21,10 @@ import { TNew, _getDecoratorMetadata } from '../../entt/internals';
  * @Serializable() decorator, configures property serialization behavior
  * @param alias (Optional) Configures property getter
  * @param serialize (Optional) Configures custom serialization mapper.
- *   Function ((target: any, value: any) => any) will be called (with a reference to the entire EnTT instance
+ *   Function ((value: any, target: any) => any) will be called (with a reference to the entire EnTT instance
  *   and the property value being fetched) and it's returned value will be used as serialized value.
  * @param deserialize (Optional) Configures custom de-serialization mapper,
- *   Function ((target: any, value: any) => any)  will be called (with a reference to the entire EnTT instance
+ *   Function ((value: any, target: any) => any)  will be called (with a reference to the entire EnTT instance
  *   and the property value being set) and it's returned value will be used  as de-serialized value.
  * @param cast (Optional) Configures how serialized value is cast before being set. Supported shapes are:
  * - { cast: MyEnTTClass }, will cast property value as instance of MyEnTTClass
@@ -36,15 +36,15 @@ import { TNew, _getDecoratorMetadata } from '../../entt/internals';
  */
 export function Serializable({
   alias = undefined as string,
-  serialize = undefined as ((target: any, value: any) => any) | boolean,
-  deserialize = undefined as ((target: any, value: any) => any) | boolean,
+  serialize = undefined as ((value: any, target: any) => any) | boolean,
+  deserialize = undefined as ((value: any, target: any) => any) | boolean,
   cast = undefined as TNew<any> | TNew<any>[] | Record<any, TNew<any>>,
 } = {}) {
   // Set defaults
   const defaults = {
     alias: undefined as string,
-    serialize: true as ((target: any, value: any) => any) | boolean,
-    deserialize: true as ((target: any, value: any) => any) | boolean,
+    serialize: true as ((value: any, target: any) => any) | boolean,
+    deserialize: true as ((value: any, target: any) => any) | boolean,
     cast: undefined as TNew<any> | TNew<any>[] | Record<any, TNew<any>>,
   };