+++ src/lib/decorators/validate/index.ts
@@ -22,16 +22,19 @@ import { _getDecoratorMetadata } from '../../entt/internals';
  * @Validate() decorator, configures property validation behavior
  * @param type Type to validate against
  * @param provider Validation provider, supports:
- * - Custom validation function of following shape: (obj, value) => Error[] | Error | string | boolean
+ * - Custom validation function of following shape: (value, obj) => Error[] | Error | string | boolean
  * - joi:         https://www.npmjs.com/package/joi
  * - joi-browser: https://www.npmjs.com/package/joi-browser
  * - yup:         https://www.npmjs.com/package/yup
  */
-export function Validate({ type = undefined as _primitiveTypeName, provider = undefined as any } = {}) {
+export function Validate({
+  type = undefined as _primitiveTypeName,
+  provider = undefined as ((value: any, obj: any) => Error[] | Error | string | boolean) | any,
+} = {}) {
   // Set defaults
   const defaults = {
     type: undefined as _primitiveTypeName,
-    provider: undefined as any,
+    provider: undefined as ((value: any, obj: any) => Error[] | Error | string | boolean) | any,
   };
 
   // Return decorator